Introduced modular version for converter between restrictedJATs and TextureJATS
