+++ src/converter/util/replaceWith.js
@@ -1,0 +1,6 @@
+export default function replaceWith(el, els) {
+  const parent = el.parentNode
+  const next = el.nextSibling
+  els.forEach(_el => parent.insertBefore(_el, next))
+  el.remove()
+}
